+++ src/ytsimpledownloader/__init__.py
@@ -1,3 +1,3 @@
 """Simple YouTube downloader."""
 
-__version__ = "0.6.0"
+__version__ = "0.6.1"